+++ sharding/nocopy.go
@@ -1,0 +1,10 @@
+package sharding
+
+// nolint: unused
+type noCopy struct{}
+
+// nolint: unused
+func (*noCopy) Lock() {}
+
+// nolint: unused
+func (*noCopy) Unlock() {}