+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - added `BitBool` for mysql bit type (#11)
+- added `sharding` feature (#12)
+
+### Fixed
+- fixed parameterized placeholder for postgresql(#12)
 
 ## [1.1.0] - 2024-02-13
 ### Added