+++ CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0]
 ### Added
 - added `BitBool` for mysql bit type (#11)
 - added `sharding` feature (#12)
-- added `On` on `DB` to enable sharding feature(#13)
+- added `On` on `DB` to enable AutoSharding feature(#13)
+- added `On` on `SQLBuilder` to enable AutoRotation feature(#13)
 
 ### Fixed
 - fixed parameterized placeholder for postgresql(#12)